Si hay un ingrediente extra entonces se cobra un 0.50 más por default pero tambien se puede agregar un precio. y cambio de int a double en metodos de pago

+++ identidades cpp/PagoEfectivo.cpp
@@ -4,10 +4,10 @@
 
 #include "../identidades h/PagoEfectivo.h"
 
-PagoEfectivo::PagoEfectivo(int monto):monto(monto) {
+PagoEfectivo::PagoEfectivo(double monto):monto(monto) {
 }
 
-bool PagoEfectivo::procesarPago(int montoPagar) {
+bool PagoEfectivo::procesarPago(double montoPagar) {
     if (monto < montoPagar) {
         cout<<"Monto insuficiente"<<endl;
         return false;
